fix: suspending computer always reset timers
Decided to simplify some of the logic to have any suspension cause a reset in the timers. This means that we don't care how long the computer was suspended for, we consider the user is rested. While this could be wrong, it's a good approximation as we don't expect the computer to suspend more quickly than the resting time. Still, we may update this in the future (just add the suspended time as a rest time).

This update will also improve some code. I removed the unnecessary model for the trayicon, some properties that had no necessity, etc. This is mostly because debugging this issue took me way longer than I expected it to take, which means the code was too complex for its own good.

Finally, I did some additional refactoring and upgrading of the code to use the new .net syntax.